Added check for button status pre-click
This would have the added advantage of making sure the page settles
before doing the click. In the automatic wait it has been observed that
although the page load (page ready) is complete. But the rendering
appears not to be complete as the initial click has a short execution of
the automatic wait. Thus checking both verifies we are in the proper
state before starting and that the page has "settled".